conferences-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
email-validator==2.2.0
aio-pika==9.4.3
